# specs/001-auth-profile-management/contracts/auth-profile.openapi.yaml
openapi: 3.0.3
info:
  title: CogniLink Auth & Profile API
  version: "1.0.0"
  description: >
    Contrato dos endpoints de autenticacao e gestao de perfil do CogniLink
    (feature 001-auth-profile-management). Todos os endpoints usam HTTPS.
    Endpoints marcados com "security: bearerAuth" exigem access token JWT
    valido e resolvem o usuario proprietario a partir das claims do token.
servers:
  - url: /api
security: []
paths:
  /auth/register:
    post:
      summary: Cadastrar novo usuario
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Conta criada; sessao iniciada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400':
          description: Dados invalidos (e-mail mal formatado, senha fraca).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: >
            Conflito de cadastro (ex.: e-mail ja em uso). Mensagem generica,
            sem confirmar a existencia de conta associada ao e-mail (FR-010).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      summary: Autenticar usuario
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: >
            Credenciais invalidas. Resposta identica para e-mail inexistente
            ou senha incorreta (FR-011).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      summary: Renovar sessao via refresh token (rotacao)
      operationId: refreshSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Nova sessao emitida; refresh token anterior invalidado (FR-004).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Refresh token expirado, revogado ou ja utilizado (FR-005).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      summary: Encerrar sessao (revogar refresh token ativo)
      operationId: logoutUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Sessao encerrada; refresh token revogado (FR-006).
        '401':
          description: Nao autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/forgot-password:
    post:
      summary: Solicitar recuperacao de senha por e-mail
      operationId: requestPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: >
            Solicitacao aceita. Resposta identica exista ou nao conta para o
            e-mail informado, para evitar enumeracao de contas.
        '400':
          description: E-mail em formato invalido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/reset-password:
    post:
      summary: Redefinir senha usando token de recuperacao
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Senha redefinida com sucesso; token invalidado.
        '400':
          description: Nova senha nao atende a politica de senha forte.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token de redefinicao invalido, expirado ou ja utilizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/change-password:
    put:
      summary: Trocar senha do usuario autenticado
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Senha alterada com sucesso.
        '400':
          description: Nova senha nao atende a politica de senha forte.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Nao autenticado, ou senha atual incorreta (FR-008).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /profile:
    get:
      summary: Consultar perfil do usuario autenticado
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil do usuario autenticado (nunca de outro usuario).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Nao autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Atualizar perfil do usuario autenticado
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Perfil atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Dados invalidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Nao autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: >
            E-mail informado ja associado a outra conta. Mensagem generica,
            sem confirmar de quem e a conta (FR-010).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          description: Deve atender a politica de senha forte definida pelo produto.

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string

    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          format: password

    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password

    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        profilePhotoUrl:
          type: string
          nullable: true
        theme:
          type: string

    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT de curta duracao.
        refreshToken:
          type: string
          description: Credencial de continuidade de longa duracao, uso unico por renovacao.
        expiresIn:
          type: integer
          description: Validade do access token em segundos.

    ProfileResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        profilePhotoUrl:
          type: string
          nullable: true
        theme:
          type: string

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Mensagem generica e segura (sem enumeracao de contas nem detalhes internos).